refactor: split files over 700 lines, separating logic from UI
Decompose 7 source files that exceeded 700 lines, following a
logic/UI separation and independent-component principle. No behavior
change; flutter analyze is clean and all tests pass.

- prompts_screen.dart 1140→672: extract prompt_dialogs.dart, prompts_io.dart
- gallery.dart 932→243: extract image_card, image_card_context_menu, gallery_file_actions
- task_queue_service.dart 908→313: move model to models/task_item.dart (re-exported),
  executors to task_executors.dart (part extension)
- token_usage_screen.dart 817→48: extract usage_stats, usage_list, usage_view_mobile/desktop
- app_state.dart 750→451: split into app_state_data / app_state_workbench part extensions
- google_genai_provider.dart 722→411: extract google_auth (re-exported), google_payload
- setup_wizard.dart 701→530: extract wizard_import.dart

Compatibility preserved via re-exports (task_item, google_auth) so dependent
files and tests need no changes. Protected notifyListeners in extensions routed
through refreshQueue()/new AppState.notify().
